Completed basic requirements for Subclass Dance Party.
We commented out Dancer.prototype.step() because we did not want to utilize the blinking pattern,
but left it in the code to show we did it properly.

Three different types of dancers are produced: Nyan Cat, Pusheen Cat, and Banana.

They line up and also shuffle around based on their nearest partner (on click in the navbar).

Easter egg included. Add 20 dancers for the first one, and then keep adding for more.
Warning: Keep the fire marshall on standby.